+++ apps/soundboardapp.js
@@ -86,10 +86,11 @@ class SoundBoardApplication extends Application {
         var volume = game.settings.get('SoundBoard', 'soundboardServerVolume');
         var collapse = totalCount > 2000;
         // TODO: Subclass mySounds, set up getData with supers
-        var players = game.users.entities.filter((el)=>el.active && !el.isGM).map((el)=>{return {name: el.name, id: el.id, isTarget:el.id==SoundBoard.targettedPlayerID?true:false};});
+        var players = (game.users.entities || game.users.contents).filter((el)=>el.active && !el.isGM).map((el)=>{return {name: el.name, id: el.id, isTarget:el.id==SoundBoard.targettedPlayerID?true:false};});
         var targettedPlayer = SoundBoard.targettedPlayerID;
         var cacheMode = SoundBoard.cacheMode;
         var macroMode = SoundBoard.macroMode;
+        var volumeMode = SoundBoard.volumeMode;
         var isExampleAudio = game.settings.get('SoundBoard', 'soundboardDirectory') == game.settings.settings.get('SoundBoard.soundboardDirectory').default;
 
         return {
@@ -104,6 +105,7 @@ class SoundBoardApplication extends Application {
             targettedPlayer,
             cacheMode,
             macroMode,
+            volumeMode,
             isExampleAudio
         };
     }