+++ helpers/soundboardsockethelper.js
@@ -7,7 +7,8 @@ class SBSocketHelper {
         STOPALL: 3,
         CACHE: 4,
         CACHECOMPLETE: 5,
-        VOLUMECHANGE: 6
+        VOLUMECHANGE: 6,
+        REQUESTMACROPLAY: 7
     }
     constructor() {
         game.socket.on(SBSocketHelper.socketName, this._onData);
@@ -16,12 +17,16 @@ class SBSocketHelper {
         if (game.user.isGM) {
             if(data.type === SBSocketHelper.SOCKETMESSAGETYPE.CACHECOMPLETE){
                 SoundBoard.audioHelper.cacheComplete(data.payload);
+            } else if (data.type == SBSocketHelper.SOCKETMESSAGETYPE.REQUESTMACROPLAY){
+                if(game.settings.get('SoundBoard', 'allowPlayersMacroRequest')){
+                    SoundBoard.playSoundByName(data.payload);
+                }
             }
         } else {
             switch (data.type) {
             case SBSocketHelper.SOCKETMESSAGETYPE.PLAY:
                 if (!data.payload.target || data.payload.target == game.userId) {
-                    SoundBoard.audioHelper.play(data.payload);
+                    SoundBoard.audioHelper.play(data.payload, data.soundExtras);
                 }
                 break;
             case SBSocketHelper.SOCKETMESSAGETYPE.STOP:
@@ -34,7 +39,7 @@ class SBSocketHelper {
                 SoundBoard.audioHelper.cache(data.payload);
                 break;
             case SBSocketHelper.SOCKETMESSAGETYPE.VOLUMECHANGE:
-                SoundBoard.audioHelper.onVolumeChange(data.payload?.volume);
+                SoundBoard.audioHelper.onVolumeChange(data.payload?.volume, data.payload?.individualVolumes);
                 break;
             default:
                 break;